+++ doc/CHANGES.rst
@@ -15,6 +15,7 @@ futur
 -----
 
 * Add option put only the primary keys
+* Fix the Front page
 
 1.0.2 (2017-10-25)
 ------------------